stats: remove the use of the global StatsOptions variable from most of the code
stats_options is global (for now) in order to decide whether a stats_counter
should be defined or not, and do that without access to GlobalConfig.

This was done probably because some counters were registered without
access to GlobalConfig. I didn't change this for now (but probably could
be changed), but at least removed the use of this global variable from
most of the code. Only a few functions use it now, whereas its
use was scattered around the file previously.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>